engine: −18 % per-op dispatch cost on the S3, and the #312 suspect table
The instruction-count model was right about the count and wrong about the
cycles per instruction: ~55 Xtensa instructions per bytecode op at ~2
cycles each. Neither cache suspect was the cause — words executed from the
flash mapping cost 0.0 % versus the DRAM heap (so #260 needs no DRAM code
cache), and putting the 15.7 KB Vm::run in IRAM bought 1.0 % for 16 KB of
the scarcest memory on the board.

What did pay, all measured on the Seengreat panel with ops/pixel unchanged
at 11.00 per loop iteration:

- Value's u16 payloads made its discriminant a u16, so every match on a
  Value — and every Option<Value> niche test Vec::pop leaves behind —
  needed `l32r 0xffff; and` first, and Xtensa has no 32-bit immediate, so
  that mask is a literal-pool LOAD re-issued at each use. All payloads
  32-bit: Add arm 34 → 22 instructions, Vm::run −1.5 KB, −12.5 %.
- Binary ops, StoreL/StoreG and ConstOp in place instead of pop+push:
  one length write instead of three, no MAX_STACK check on a shape that
  shrinks the stack. −2.3 % (and +9 % slower on x86).
- fuel and insn_start into locals: a load/compare/decrement/store and a
  store on every dispatch, for two values only the error paths read.
  Published at each return and at the two sites that can re-enter the VM.
  −1.5 %.
- Instruction fetch merges on a value, not on control flow, and naming
  opcode 0 starts the jump table's range at 0. Preamble 20 → 16
  instructions, −2.5 %.

Net: 110.3 → 90.5 cycles/op, loop iteration 5.055 → 4.147 µs; rainbow vm
−2.5 %, snake 1D −7.5 % (17→18 fps), snake 2D −11.1 % (10→11 fps), empty
render unchanged (it executes no ops). 40/40 byte-identical PPMs against
master, check-library 1495/1495.

Also: tools/opbench.mjs (the measurement), the iram-vm / iram-builtins
features that suspect 2 was tested with (off everywhere), and the
ops/px × cycles/op rule plus the host-vs-device divergence in
.claude/rules/vm-bytecode.md.

Gitea: #312

+++ crates/luxel-core/src/compile.rs
@@ -398,8 +398,8 @@ fn emit_insn(out: &mut Vec<u32>, insn: &Insn, offsets: &[u32]) {
             out.push(enc::bare(op::CONST_NUM));
             0
         }
-        Const(Value::Fun(i)) => enc::with_u16(op::CONST_FUN, *i),
-        Const(Value::Builtin(b)) => enc::with_u16(op::CONST_BUILTIN, *b),
+        Const(Value::Fun(i)) => enc::with_u16(op::CONST_FUN, *i as u16),
+        Const(Value::Builtin(b)) => enc::with_u16(op::CONST_BUILTIN, *b as u16),
         LoadG(i) => enc::with_u16(op::LOAD_G, *i),
         StoreG(i) => enc::with_u16(op::STORE_G, *i),
         LoadL(i) => enc::with_u8(op::LOAD_L, *i),
@@ -866,7 +866,7 @@ impl<'s> Compiler<'s> {
                 .map(|&(_, i, _)| i)
                 .expect("demoted implies registered");
             let g = self.ensure_global(&name, false, Span::default())?;
-            ctx.push(Insn::Const(Value::Fun(idx)));
+            ctx.push(Insn::Const(Value::Fun(idx as u32)));
             ctx.push(Insn::StoreG(g));
             ctx.push(Insn::Pop);
         }
@@ -1288,8 +1288,8 @@ impl<'s> Compiler<'s> {
                 match self.resolve(ctx, name, e.span)? {
                     Place::Local(i) => ctx.push(Insn::LoadL(i)),
                     Place::Global(i) => ctx.push(Insn::LoadG(i)),
-                    Place::Func(i) => ctx.push(Insn::Const(Value::Fun(i))),
-                    Place::Builtin(b) => ctx.push(Insn::Const(Value::Builtin(b))),
+                    Place::Func(i) => ctx.push(Insn::Const(Value::Fun(i as u32))),
+                    Place::Builtin(b) => ctx.push(Insn::Const(Value::Builtin(b as u32))),
                 }
                 Ok(())
             }
@@ -1499,7 +1499,7 @@ impl<'s> Compiler<'s> {
             }
             ExprKind::Lambda { params, body } => {
                 let idx = self.emit_lambda(params, body, e.span)?;
-                ctx.push(Insn::Const(Value::Fun(idx)));
+                ctx.push(Insn::Const(Value::Fun(idx as u32)));
                 Ok(())
             }
         }
